requirements.txt
aiohttp==3.9.5
aiohttp-json-rpc==0.13.3
aiosignal==1.3.1
apsw==3.46.1.0
attrs==24.2.0
babel==2.16.0
Brotli==1.1.0
click==8.1.7
discord.py==2.4.0
distro==1.9.0
frozenlist==1.4.1
idna==3.8
Markdown==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
orjson==3.10.7
packaging==24.1
platformdirs==4.2.2
psutil==6.0.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
rapidfuzz==3.9.6
Red-Commons==1.0.0
Red-DiscordBot==3.5.13
Red-Lavalink==0.11.0
rich==13.8.0
schema==0.7.7
six==1.16.0
typing_extensions==4.12.2
uvloop==0.20.0
yarl==1.9.4
